extract RecordingActions into its own file
