+++ src/app/RecordingList/ArchivedRecordingsList.tsx
@@ -40,7 +40,7 @@ import { Recording } from '@app/Shared/Services/Api.service';
 import { ServiceContext } from '@app/Shared/Services/Services';
 import { useSubscriptions } from '@app/utils/useSubscriptions';
 import { Button, DataListCell, DataListCheck, DataListContent, DataListItem, DataListItemCells, DataListItemRow, DataListToggle, Toolbar, ToolbarContent, ToolbarGroup, ToolbarItem } from '@patternfly/react-core';
-import { RecordingActions } from './ActiveRecordingsList';
+import { RecordingActions } from './RecordingActions';
 import { RecordingsDataTable } from './RecordingsDataTable';
 import { ReportFrame } from './ReportFrame';
 import { Observable, Subject, forkJoin } from 'rxjs';
